+++ src/ReaderT.ts
@@ -1,19 +1,23 @@
 /**
  * @since 2.0.0
  */
-import { Apply, Apply1, Apply2 } from './Apply'
-import { Chain, Chain1, Chain2 } from './Chain'
+import { Apply, Apply1, Apply2, Apply2C, Apply3, Apply3C, Apply4 } from './Apply'
+import { Chain, Chain1, Chain2, Chain2C, Chain3, Chain3C, Chain4 } from './Chain'
 import { flow, pipe } from './function'
-import { Functor, Functor1, Functor2 } from './Functor'
-import { HKT, Kind, Kind2, Kind3, URIS, URIS2, URIS3 } from './HKT'
+import { Functor, Functor1, Functor2, Functor2C, Functor3, Functor3C, Functor4 } from './Functor'
+import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
 import { Monad, Monad1, Monad2, Monad2C, Monad3 } from './Monad'
-import { Pointed, Pointed1, Pointed2 } from './Pointed'
+import { Pointed, Pointed1, Pointed2, Pointed2C, Pointed3, Pointed3C, Pointed4 } from './Pointed'
 import { Reader } from './Reader'
 
 /**
  * @since 2.10.0
  */
-export function of<F extends URIS2>(F: Pointed2<F>): <A, R, ME>(a: A) => Reader<R, Kind2<F, ME, A>>
+export function of<F extends URIS4>(F: Pointed4<F>): <A, R, S, FR, FE>(a: A) => Reader<R, Kind4<F, S, FR, FE, A>>
+export function of<F extends URIS3>(F: Pointed3<F>): <A, R, FR, FE>(a: A) => Reader<R, Kind3<F, FR, FE, A>>
+export function of<F extends URIS3, FE>(F: Pointed3C<F, FE>): <A, R, FR>(a: A) => Reader<R, Kind3<F, FR, FE, A>>
+export function of<F extends URIS2>(F: Pointed2<F>): <A, R, FE>(a: A) => Reader<R, Kind2<F, FE, A>>
+export function of<F extends URIS2, FE>(F: Pointed2C<F, FE>): <A, R>(a: A) => Reader<R, Kind2<F, FE, A>>
 export function of<F extends URIS>(F: Pointed1<F>): <A, R>(a: A) => Reader<R, Kind<F, A>>
 export function of<F>(F: Pointed<F>): <A, R>(a: A) => Reader<R, HKT<F, A>>
 export function of<F>(F: Pointed<F>): <A, R>(a: A) => Reader<R, HKT<F, A>> {
@@ -23,9 +27,21 @@ export function of<F>(F: Pointed<F>): <A, R>(a: A) => Reader<R, HKT<F, A>> {
 /**
  * @since 2.10.0
  */
+export function map<F extends URIS4>(
+  F: Functor4<F>
+): <A, B>(f: (a: A) => B) => <R, S, FR, FE>(fa: Reader<R, Kind4<F, S, FR, FE, A>>) => Reader<R, Kind4<F, S, FR, FE, B>>
+export function map<F extends URIS3>(
+  F: Functor3<F>
+): <A, B>(f: (a: A) => B) => <R, FR, FE>(fa: Reader<R, Kind3<F, FR, FE, A>>) => Reader<R, Kind3<F, FR, FE, B>>
+export function map<F extends URIS3, FE>(
+  F: Functor3C<F, FE>
+): <A, B>(f: (a: A) => B) => <R, FR>(fa: Reader<R, Kind3<F, FR, FE, A>>) => Reader<R, Kind3<F, FR, FE, B>>
 export function map<F extends URIS2>(
   F: Functor2<F>
 ): <A, B>(f: (a: A) => B) => <R, FE>(fa: Reader<R, Kind2<F, FE, A>>) => Reader<R, Kind2<F, FE, B>>
+export function map<F extends URIS2, FE>(
+  F: Functor2C<F, FE>
+): <A, B>(f: (a: A) => B) => <R>(fa: Reader<R, Kind2<F, FE, A>>) => Reader<R, Kind2<F, FE, B>>
 export function map<F extends URIS>(
   F: Functor1<F>
 ): <A, B>(f: (a: A) => B) => <R>(fa: Reader<R, Kind<F, A>>) => Reader<R, Kind<F, B>>
@@ -37,11 +53,31 @@ export function map<F>(F: Functor<F>): <A, B>(f: (a: A) => B) => <R>(fa: Reader<
 /**
  * @since 2.10.0
  */
+export function ap<F extends URIS4>(
+  F: Apply4<F>
+): <R, S, FR, FE, A>(
+  fa: Reader<R, Kind4<F, S, FR, FE, A>>
+) => <B>(fab: Reader<R, Kind4<F, S, FR, FE, (a: A) => B>>) => Reader<R, Kind4<F, S, FR, FE, B>>
+export function ap<F extends URIS3>(
+  F: Apply3<F>
+): <R, FR, FE, A>(
+  fa: Reader<R, Kind3<F, FR, FE, A>>
+) => <B>(fab: Reader<R, Kind3<F, FR, FE, (a: A) => B>>) => Reader<R, Kind3<F, FR, FE, B>>
+export function ap<F extends URIS3, FE>(
+  F: Apply3C<F, FE>
+): <R, FR, A>(
+  fa: Reader<R, Kind3<F, FR, FE, A>>
+) => <B>(fab: Reader<R, Kind3<F, FR, FE, (a: A) => B>>) => Reader<R, Kind3<F, FR, FE, B>>
 export function ap<F extends URIS2>(
   F: Apply2<F>
-): <R, E, A>(
-  fa: Reader<R, Kind2<F, E, A>>
-) => <B>(fab: Reader<R, Kind2<F, E, (a: A) => B>>) => Reader<R, Kind2<F, E, B>>
+): <R, FE, A>(
+  fa: Reader<R, Kind2<F, FE, A>>
+) => <B>(fab: Reader<R, Kind2<F, FE, (a: A) => B>>) => Reader<R, Kind2<F, FE, B>>
+export function ap<F extends URIS2, FE>(
+  F: Apply2C<F, FE>
+): <R, A>(
+  fa: Reader<R, Kind2<F, FE, A>>
+) => <B>(fab: Reader<R, Kind2<F, FE, (a: A) => B>>) => Reader<R, Kind2<F, FE, B>>
 export function ap<F extends URIS>(
   F: Apply1<F>
 ): <R, A>(fa: Reader<R, Kind<F, A>>) => <B>(fab: Reader<R, Kind<F, (a: A) => B>>) => Reader<R, Kind<F, B>>
@@ -57,9 +93,29 @@ export function ap<F>(
 /**
  * @since 2.10.0
  */
+export function chain<M extends URIS4>(
+  M: Chain4<M>
+): <A, R, S, FR, FE, B>(
+  f: (a: A) => Reader<R, Kind4<M, S, FR, FE, B>>
+) => (ma: Reader<R, Kind4<M, S, FR, FE, A>>) => Reader<R, Kind4<M, S, FR, FE, B>>
+export function chain<M extends URIS3>(
+  M: Chain3<M>
+): <A, R, FR, FE, B>(
+  f: (a: A) => Reader<R, Kind3<M, FR, FE, B>>
+) => (ma: Reader<R, Kind3<M, FR, FE, A>>) => Reader<R, Kind3<M, FR, FE, B>>
+export function chain<M extends URIS3, FE>(
+  M: Chain3C<M, FE>
+): <A, R, FR, B>(
+  f: (a: A) => Reader<R, Kind3<M, FR, FE, B>>
+) => (ma: Reader<R, Kind3<M, FR, FE, A>>) => Reader<R, Kind3<M, FR, FE, B>>
 export function chain<M extends URIS2>(
   M: Chain2<M>
-): <A, R, E, B>(f: (a: A) => Reader<R, Kind2<M, E, B>>) => (ma: Reader<R, Kind2<M, E, A>>) => Reader<R, Kind2<M, E, B>>
+): <A, R, FE, B>(
+  f: (a: A) => Reader<R, Kind2<M, FE, B>>
+) => (ma: Reader<R, Kind2<M, FE, A>>) => Reader<R, Kind2<M, FE, B>>
+export function chain<M extends URIS2, FE>(
+  M: Chain2C<M, FE>
+): <A, R, B>(f: (a: A) => Reader<R, Kind2<M, FE, B>>) => (ma: Reader<R, Kind2<M, FE, A>>) => Reader<R, Kind2<M, FE, B>>
 export function chain<M extends URIS>(
   M: Chain1<M>
 ): <A, R, B>(f: (a: A) => Reader<R, Kind<M, B>>) => (ma: Reader<R, Kind<M, A>>) => Reader<R, Kind<M, B>>
@@ -75,7 +131,11 @@ export function chain<M>(
 /**
  * @since 2.10.0
  */
-export function ask<F extends URIS2>(F: Pointed2<F>): <R, E>() => Reader<R, Kind2<F, E, R>>
+export function ask<F extends URIS4>(F: Pointed4<F>): <R, S, FR, FE>() => Reader<R, Kind4<F, S, FR, FE, R>>
+export function ask<F extends URIS3>(F: Pointed3<F>): <R, FR, FE>() => Reader<R, Kind3<F, FR, FE, R>>
+export function ask<F extends URIS3, FE>(F: Pointed3C<F, FE>): <R, FR>() => Reader<R, Kind3<F, FR, FE, R>>
+export function ask<F extends URIS2>(F: Pointed2<F>): <R, FE>() => Reader<R, Kind2<F, FE, R>>
+export function ask<F extends URIS2, FE>(F: Pointed2C<F, FE>): <R>() => Reader<R, Kind2<F, FE, R>>
 export function ask<F extends URIS>(F: Pointed1<F>): <R>() => Reader<R, Kind<F, R>>
 export function ask<F>(F: Pointed<F>): <R>() => Reader<R, HKT<F, R>>
 export function ask<F>(F: Pointed<F>): <R>() => Reader<R, HKT<F, R>> {
@@ -85,7 +145,15 @@ export function ask<F>(F: Pointed<F>): <R>() => Reader<R, HKT<F, R>> {
 /**
  * @since 2.10.0
  */
-export function asks<F extends URIS2>(F: Pointed2<F>): <R, A, E>(f: (r: R) => A) => Reader<R, Kind2<F, E, A>>
+export function asks<F extends URIS4>(
+  F: Pointed4<F>
+): <R, A, S, FR, FE>(f: (r: R) => A) => Reader<R, Kind4<F, S, FR, FE, A>>
+export function asks<F extends URIS3>(F: Pointed3<F>): <R, A, FR, FE>(f: (r: R) => A) => Reader<R, Kind3<F, FR, FE, A>>
+export function asks<F extends URIS3, FE>(
+  F: Pointed3C<F, FE>
+): <R, A, FR>(f: (r: R) => A) => Reader<R, Kind3<F, FR, FE, A>>
+export function asks<F extends URIS2>(F: Pointed2<F>): <R, A, FE>(f: (r: R) => A) => Reader<R, Kind2<F, FE, A>>
+export function asks<F extends URIS2, FE>(F: Pointed2C<F, FE>): <R, A>(f: (r: R) => A) => Reader<R, Kind2<F, FE, A>>
 export function asks<F extends URIS>(F: Pointed1<F>): <R, A>(f: (r: R) => A) => Reader<R, Kind<F, A>>
 export function asks<F>(F: Pointed<F>): <R, A>(f: (r: R) => A) => Reader<R, HKT<F, A>>
 export function asks<F>(F: Pointed<F>): <R, A>(f: (r: R) => A) => Reader<R, HKT<F, A>> {
@@ -95,7 +163,19 @@ export function asks<F>(F: Pointed<F>): <R, A>(f: (r: R) => A) => Reader<R, HKT<
 /**
  * @since 2.10.0
  */
-export function fromReader<F extends URIS2>(F: Pointed2<F>): <R, A, E>(ma: Reader<R, A>) => Reader<R, Kind2<F, E, A>>
+export function fromReader<F extends URIS4>(
+  F: Pointed4<F>
+): <R, A, S, FR, FE>(ma: Reader<R, A>) => Reader<R, Kind4<F, S, FR, FE, A>>
+export function fromReader<F extends URIS3>(
+  F: Pointed3<F>
+): <R, A, FR, FE>(ma: Reader<R, A>) => Reader<R, Kind3<F, FR, FE, A>>
+export function fromReader<F extends URIS3, FE>(
+  F: Pointed3C<F, FE>
+): <R, A, FR>(ma: Reader<R, A>) => Reader<R, Kind3<F, FR, FE, A>>
+export function fromReader<F extends URIS2>(F: Pointed2<F>): <R, A, FE>(ma: Reader<R, A>) => Reader<R, Kind2<F, FE, A>>
+export function fromReader<F extends URIS2, FE>(
+  F: Pointed2C<F, FE>
+): <R, A>(ma: Reader<R, A>) => Reader<R, Kind2<F, FE, A>>
 export function fromReader<F extends URIS>(F: Pointed1<F>): <R, A>(ma: Reader<R, A>) => Reader<R, Kind<F, A>>
 export function fromReader<F>(F: Pointed<F>): <R, A>(ma: Reader<R, A>) => Reader<R, HKT<F, A>>
 export function fromReader<F>(F: Pointed<F>): <R, A>(ma: Reader<R, A>) => Reader<R, HKT<F, A>> {