+++ src/StateT.ts
@@ -1,12 +1,12 @@
 /**
  * @since 2.0.0
  */
-import { Chain, Chain1, Chain2, Chain3 } from './Chain'
+import { Chain, Chain1, Chain2, Chain2C, Chain3, Chain3C } from './Chain'
 import { Endomorphism, pipe } from './function'
-import { Functor, Functor1, Functor2, Functor3 } from './Functor'
+import { Functor, Functor1, Functor2, Functor2C, Functor3, Functor3C } from './Functor'
 import { HKT, Kind, Kind2, Kind3, URIS, URIS2, URIS3 } from './HKT'
 import { Monad, Monad1, Monad2, Monad2C, Monad3, Monad3C } from './Monad'
-import { Pointed, Pointed1, Pointed2, Pointed3 } from './Pointed'
+import { Pointed, Pointed1, Pointed2, Pointed2C, Pointed3, Pointed3C } from './Pointed'
 import { snd } from './ReadonlyTuple'
 import { State } from './State'
 
@@ -45,7 +45,9 @@ export interface StateT3<M extends URIS3, S, R, E, A> {
  * @since 2.10.0
  */
 export function of<F extends URIS3>(F: Pointed3<F>): <A, S, R, E>(a: A) => StateT3<F, S, R, E, A>
+export function of<F extends URIS3, E>(F: Pointed3C<F, E>): <A, S, R>(a: A) => StateT3<F, S, R, E, A>
 export function of<F extends URIS2>(F: Pointed2<F>): <A, S, E>(a: A) => StateT2<F, S, E, A>
+export function of<F extends URIS2, E>(F: Pointed2C<F, E>): <A, S>(a: A) => StateT2<F, S, E, A>
 export function of<F extends URIS>(F: Pointed1<F>): <A, S>(a: A) => StateT1<F, S, A>
 export function of<F>(F: Pointed<F>): <A, S>(a: A) => StateT<F, S, A>
 export function of<F>(F: Pointed<F>): <A, S>(a: A) => StateT<F, S, A> {
@@ -58,9 +60,15 @@ export function of<F>(F: Pointed<F>): <A, S>(a: A) => StateT<F, S, A> {
 export function map<F extends URIS3>(
   F: Functor3<F>
 ): <A, B>(f: (a: A) => B) => <S, R, E>(fa: StateT3<F, S, R, E, A>) => StateT3<F, S, R, E, B>
+export function map<F extends URIS3, E>(
+  F: Functor3C<F, E>
+): <A, B>(f: (a: A) => B) => <S, R>(fa: StateT3<F, S, R, E, A>) => StateT3<F, S, R, E, B>
 export function map<F extends URIS2>(
   F: Functor2<F>
 ): <A, B>(f: (a: A) => B) => <S, E>(fa: StateT2<F, S, E, A>) => StateT2<F, S, E, B>
+export function map<F extends URIS2, E>(
+  F: Functor2C<F, E>
+): <A, B>(f: (a: A) => B) => <S>(fa: StateT2<F, S, E, A>) => StateT2<F, S, E, B>
 export function map<F extends URIS>(
   F: Functor1<F>
 ): <A, B>(f: (a: A) => B) => <S>(fa: StateT1<F, S, A>) => StateT1<F, S, B>
@@ -75,9 +83,15 @@ export function map<F>(F: Functor<F>): <A, B>(f: (a: A) => B) => <S>(fa: StateT<
 export function ap<M extends URIS3>(
   M: Chain3<M>
 ): <S, R, E, A>(fa: StateT3<M, S, R, E, A>) => <B>(fab: StateT3<M, S, R, E, (a: A) => B>) => StateT3<M, S, R, E, B>
+export function ap<M extends URIS3, E>(
+  M: Chain3C<M, E>
+): <S, R, A>(fa: StateT3<M, S, R, E, A>) => <B>(fab: StateT3<M, S, R, E, (a: A) => B>) => StateT3<M, S, R, E, B>
 export function ap<M extends URIS2>(
   M: Chain2<M>
 ): <S, E, A>(fa: StateT2<M, S, E, A>) => <B>(fab: StateT2<M, S, E, (a: A) => B>) => StateT2<M, S, E, B>
+export function ap<M extends URIS2, E>(
+  M: Chain2C<M, E>
+): <S, A>(fa: StateT2<M, S, E, A>) => <B>(fab: StateT2<M, S, E, (a: A) => B>) => StateT2<M, S, E, B>
 export function ap<M extends URIS>(
   M: Chain1<M>
 ): <S, A>(fa: StateT1<M, S, A>) => <B>(fab: StateT1<M, S, (a: A) => B>) => StateT1<M, S, B>
@@ -96,9 +110,15 @@ export function ap<M>(
 export function chain<M extends URIS3>(
   M: Chain3<M>
 ): <A, S, R, E, B>(f: (a: A) => StateT3<M, S, R, E, B>) => (ma: StateT3<M, S, R, E, A>) => StateT3<M, S, R, E, B>
+export function chain<M extends URIS3, E>(
+  M: Chain3C<M, E>
+): <A, S, R, B>(f: (a: A) => StateT3<M, S, R, E, B>) => (ma: StateT3<M, S, R, E, A>) => StateT3<M, S, R, E, B>
 export function chain<M extends URIS2>(
   M: Chain2<M>
 ): <A, S, E, B>(f: (a: A) => StateT2<M, S, E, B>) => (ma: StateT2<M, S, E, A>) => StateT2<M, S, E, B>
+export function chain<M extends URIS2, E>(
+  M: Chain2C<M, E>
+): <A, S, B>(f: (a: A) => StateT2<M, S, E, B>) => (ma: StateT2<M, S, E, A>) => StateT2<M, S, E, B>
 export function chain<M extends URIS>(
   M: Chain1<M>
 ): <A, S, B>(f: (a: A) => StateT1<M, S, B>) => (ma: StateT1<M, S, A>) => StateT1<M, S, B>
@@ -115,7 +135,9 @@ export function chain<M>(
  * @since 2.10.0
  */
 export function get<F extends URIS3>(F: Pointed3<F>): <S, R, E>() => StateT3<F, S, R, E, S>
+export function get<F extends URIS3, E>(F: Pointed3C<F, E>): <S, R>() => StateT3<F, S, R, E, S>
 export function get<F extends URIS2>(F: Pointed2<F>): <S, E>() => StateT2<F, S, E, S>
+export function get<F extends URIS2, E>(F: Pointed2C<F, E>): <S>() => StateT2<F, S, E, S>
 export function get<F extends URIS>(F: Pointed1<F>): <S>() => StateT1<F, S, S>
 export function get<F>(F: Pointed<F>): <S>() => StateT<F, S, S>
 export function get<F>(F: Pointed<F>): <S>() => StateT<F, S, S> {
@@ -126,7 +148,9 @@ export function get<F>(F: Pointed<F>): <S>() => StateT<F, S, S> {
  * @since 2.10.0
  */
 export function put<F extends URIS3>(F: Pointed3<F>): <S, R, E>(s: S) => StateT3<F, S, R, E, void>
+export function put<F extends URIS3, E>(F: Pointed3C<F, E>): <S, R>(s: S) => StateT3<F, S, R, E, void>
 export function put<F extends URIS2>(F: Pointed2<F>): <S, E>(s: S) => StateT2<F, S, E, void>
+export function put<F extends URIS2, E>(F: Pointed2C<F, E>): <S>(s: S) => StateT2<F, S, E, void>
 export function put<F extends URIS>(F: Pointed1<F>): <S>(s: S) => StateT1<F, S, void>
 export function put<F>(F: Pointed<F>): <S>(s: S) => StateT<F, S, void>
 export function put<F>(F: Pointed<F>): <S>(s: S) => StateT<F, S, void> {
@@ -137,7 +161,9 @@ export function put<F>(F: Pointed<F>): <S>(s: S) => StateT<F, S, void> {
  * @since 2.10.0
  */
 export function modify<F extends URIS3>(F: Pointed3<F>): <S, R, E>(f: Endomorphism<S>) => StateT3<F, S, R, E, void>
+export function modify<F extends URIS3, E>(F: Pointed3C<F, E>): <S, R>(f: Endomorphism<S>) => StateT3<F, S, R, E, void>
 export function modify<F extends URIS2>(F: Pointed2<F>): <S, E>(f: Endomorphism<S>) => StateT2<F, S, E, void>
+export function modify<F extends URIS2, E>(F: Pointed2C<F, E>): <S>(f: Endomorphism<S>) => StateT2<F, S, E, void>
 export function modify<F extends URIS>(F: Pointed1<F>): <S>(f: Endomorphism<S>) => StateT1<F, S, void>
 export function modify<F>(F: Pointed<F>): <S>(f: Endomorphism<S>) => StateT<F, S, void>
 export function modify<F>(F: Pointed<F>): <S>(f: Endomorphism<S>) => StateT<F, S, void> {
@@ -148,7 +174,9 @@ export function modify<F>(F: Pointed<F>): <S>(f: Endomorphism<S>) => StateT<F, S
  * @since 2.10.0
  */
 export function gets<F extends URIS3>(F: Pointed3<F>): <S, A, R, E>(f: (s: S) => A) => StateT3<F, S, R, E, A>
+export function gets<F extends URIS3, E>(F: Pointed3C<F, E>): <S, A, R>(f: (s: S) => A) => StateT3<F, S, R, E, A>
 export function gets<F extends URIS2>(F: Pointed2<F>): <S, A, E>(f: (s: S) => A) => StateT2<F, S, E, A>
+export function gets<F extends URIS2, E>(F: Pointed2C<F, E>): <S, A>(f: (s: S) => A) => StateT2<F, S, E, A>
 export function gets<F extends URIS>(F: Pointed1<F>): <S, A>(f: (s: S) => A) => StateT1<F, S, A>
 export function gets<F>(F: Pointed<F>): <S, A>(f: (s: S) => A) => StateT<F, S, A>
 export function gets<F>(F: Pointed<F>): <S, A>(f: (s: S) => A) => StateT<F, S, A> {
@@ -159,7 +187,9 @@ export function gets<F>(F: Pointed<F>): <S, A>(f: (s: S) => A) => StateT<F, S, A
  * @since 2.10.0
  */
 export function fromState<F extends URIS3>(F: Pointed3<F>): <S, A, R, E>(sa: State<S, A>) => StateT3<F, S, R, E, A>
+export function fromState<F extends URIS3, E>(F: Pointed3C<F, E>): <S, A, R>(sa: State<S, A>) => StateT3<F, S, R, E, A>
 export function fromState<F extends URIS2>(F: Pointed2<F>): <S, A, E>(sa: State<S, A>) => StateT2<F, S, E, A>
+export function fromState<F extends URIS2, E>(F: Pointed2C<F, E>): <S, A>(sa: State<S, A>) => StateT2<F, S, E, A>
 export function fromState<F extends URIS>(F: Pointed1<F>): <S, A>(sa: State<S, A>) => StateT1<F, S, A>
 export function fromState<F>(F: Pointed<F>): <S, A>(sa: State<S, A>) => StateT<F, S, A>
 export function fromState<F>(F: Pointed<F>): <S, A>(sa: State<S, A>) => StateT<F, S, A> {
@@ -170,7 +200,11 @@ export function fromState<F>(F: Pointed<F>): <S, A>(sa: State<S, A>) => StateT<F
  * @since 2.10.0
  */
 export function fromF<F extends URIS3>(F: Functor3<F>): <R, E, A, S>(ma: Kind3<F, R, E, A>) => StateT3<F, S, R, E, A>
+export function fromF<F extends URIS3, E>(
+  F: Functor3C<F, E>
+): <R, A, S>(ma: Kind3<F, R, E, A>) => StateT3<F, S, R, E, A>
 export function fromF<F extends URIS2>(F: Functor2<F>): <E, A, S>(ma: Kind2<F, E, A>) => StateT2<F, S, E, A>
+export function fromF<F extends URIS2, E>(F: Functor2C<F, E>): <A, S>(ma: Kind2<F, E, A>) => StateT2<F, S, E, A>
 export function fromF<F extends URIS>(F: Functor1<F>): <A, S>(ma: Kind<F, A>) => StateT<F, S, A>
 export function fromF<F>(F: Functor<F>): <A, S>(ma: HKT<F, A>) => StateT<F, S, A>
 export function fromF<F>(F: Functor<F>): <A, S>(ma: HKT<F, A>) => StateT<F, S, A> {
@@ -183,9 +217,15 @@ export function fromF<F>(F: Functor<F>): <A, S>(ma: HKT<F, A>) => StateT<F, S, A
 export function evaluate<F extends URIS3>(
   F: Functor3<F>
 ): <S>(s: S) => <R, E, A>(ma: StateT3<F, S, R, E, A>) => Kind3<F, R, E, A>
+export function evaluate<F extends URIS3, E>(
+  F: Functor3C<F, E>
+): <S>(s: S) => <R, A>(ma: StateT3<F, S, R, E, A>) => Kind3<F, R, E, A>
 export function evaluate<F extends URIS2>(
   F: Functor2<F>
 ): <S>(s: S) => <E, A>(ma: StateT2<F, S, E, A>) => Kind2<F, E, A>
+export function evaluate<F extends URIS2, E>(
+  F: Functor2C<F, E>
+): <S>(s: S) => <A>(ma: StateT2<F, S, E, A>) => Kind2<F, E, A>
 export function evaluate<F extends URIS>(F: Functor1<F>): <S>(s: S) => <A>(ma: StateT1<F, S, A>) => Kind<F, A>
 export function evaluate<F>(F: Functor<F>): <S>(s: S) => <A>(ma: StateT<F, S, A>) => HKT<F, A>
 export function evaluate<F>(F: Functor<F>): <S>(s: S) => <A>(ma: StateT<F, S, A>) => HKT<F, A> {
@@ -198,7 +238,13 @@ export function evaluate<F>(F: Functor<F>): <S>(s: S) => <A>(ma: StateT<F, S, A>
 export function execute<F extends URIS3>(
   F: Functor3<F>
 ): <S>(s: S) => <R, E, A>(ma: StateT3<F, S, R, E, A>) => Kind3<F, R, E, S>
+export function execute<F extends URIS3, E>(
+  F: Functor3C<F, E>
+): <S>(s: S) => <R, A>(ma: StateT3<F, S, R, E, A>) => Kind3<F, R, E, S>
 export function execute<F extends URIS2>(F: Functor2<F>): <S>(s: S) => <E, A>(ma: StateT2<F, S, E, A>) => Kind2<F, E, S>
+export function execute<F extends URIS2, E>(
+  F: Functor2C<F, E>
+): <S>(s: S) => <A>(ma: StateT2<F, S, E, A>) => Kind2<F, E, S>
 export function execute<F extends URIS>(F: Functor1<F>): <S>(s: S) => <A>(ma: StateT1<F, S, A>) => Kind<F, S>
 export function execute<F>(F: Functor<F>): <S>(s: S) => <A>(ma: StateT<F, S, A>) => HKT<F, S>
 export function execute<F>(F: Functor<F>): <S>(s: S) => <A>(ma: StateT<F, S, A>) => HKT<F, S> {